Move the test case
